requirements.txt
GPy==0.6.0
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.0
certifi==2015.04.28
funcsigs==0.4
functools32==3.2.3-2
ipython==3.2.1
jsonschema==2.5.1
matplotlib==1.4.3
mistune==0.7
mock==1.3.0
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
path.py==7.4
pbr==1.5.0
ptyprocess==0.5
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.7.0
requests==2.7.0
scikit-learn==0.16.1
scipy==0.15.1
seaborn==0.6.0
six==1.9.0
terminado==0.5
tornado==4.2.1
tqdm==1.0
wsgiref==0.1.2
